Roadblock hit with golang color library
So the original plan was to color a text document with different
colors depending on attributes of the words.  Except that color
pulls a new line when ever it colors a word, and quickly skimming
through the go color library it does not have a flag that removes
that.  I will need to come up with a new other function.